MVPにする途中
