+++ app/src/main/java/application/android/marshi/papercrane/di/component/AppComponent.java
@@ -6,6 +6,7 @@ import application.android.marshi.papercrane.activity.MainActivity;
 import application.android.marshi.papercrane.di.module.AppModule;
 import application.android.marshi.papercrane.di.module.DatabaseModule;
 import application.android.marshi.papercrane.di.module.RepositoryModule;
+import application.android.marshi.papercrane.di.module.UseCaseModule;
 import application.android.marshi.papercrane.fragment.TweetListFragment;
 import dagger.Component;
 
@@ -15,7 +16,7 @@ import javax.inject.Singleton;
  * @author marshi on 2016/04/17.
  */
 @Singleton
-@Component(modules = {RepositoryModule.class, AppModule.class, DatabaseModule.class})
+@Component(modules = {RepositoryModule.class, AppModule.class, UseCaseModule.class, DatabaseModule.class})
 public interface AppComponent {
 
 	void inject(MainActivity mainActivity);
